--- jdaviz/configs/imviz/tests/conftest.py
@@ -1,151 +1,0 @@
-# -*- coding: utf-8 -*-
-import os
-import pytest
-
-# This conftest file programmatically skips a set of known-failing
-# cubeviz tests. It is intended as a short-term measure to allow the
-# rest of the test-suite to run while these tests are being fixed.
-#
-# The list below contains nodeid fragments for tests to skip. We use
-# substring matching against collected item.nodeid so that variations
-# in exact nodeid formatting (paths/markers) still match.
-
-SKIP_TEST_NODEIDS = [
-    "test_catalogs.py::test_from_file_parsing",
-    "test_catalogs.py::test_offline_ecsv_catalog",
-    "test_catalogs.py::test_zoom_to_selected", 
-    "test_catalogs.py::test_select_tool",
-    "test_compass.py::test_user_api",
-    "test_footprints.py::test_user_api",
-    "test_footprints.py::test_api_after_linking",
-    "test_footprints.py::test_footprint_updates_on_rotation",
-    "test_footprints.py::test_footprint_select",
-    "test_helper.py::test_create_new_viewer",
-    "test_helper.py::test_temporary_imviz_current_app",
-    "test_line_profile_xy.py::test_line_profile_with_nan",
-    "test_linking.py::test_imviz_no_data",
-    "test_parser.py::TestParseImage::test_parse_numpy_array_1d_2d",
-    "test_parser.py::TestParseImage::test_parse_numpy_array_3d",
-    "test_parser.py::TestParseImage::test_parse_numpy_array_3d_too_many",
-    "test_parser.py::TestParseImage::test_parse_numpy_array_4d",
-    "test_parser.py::TestParseImage::test_parse_rgba",
-    "test_parser.py::TestParseImage::test_filelist",
-    "test_parser.py::test_roman_parser",
-    "test_parser_roman.py::test_roman_wfi_ext_options",
-    "test_regions.py::TestLoadRegionsFromFile::test_ds9_load_one_good_one_bad",
-    "test_simple_aper_phot.py::test_annulus_background",
-    "test_simple_aper_phot.py::test_fit_radial_profile_with_nan",
-    "test_tools.py::test_panzoom_click_center_linking",
-    "test_tools.py::test_blink",
-    "test_tools.py::test_compass_open_while_load",
-    "test_tools.py::test_tool_visibility",
-    "test_viewers.py::test_create_destroy_viewer",
-    "test_viewers.py::test_create_viewer_align_by_wcs",
-    "test_viewers.py::test_align_by_wcs_create_viewer",
-    "test_viewers.py::test_get_viewer_created",
-    "test_viewers.py::test_destroy_viewer_invalid",
-    "test_viewers.py::test_destroy_viewer_with_subset",
-    "test_viewers.py::test_zoom_center_radius_init",
-    "test_viewers.py::test_catalog_in_image_viewer",
-    "test_viewers.py::test_get_viewport_sky_region_wcs",
-    "test_viewers.py::test_get_viewport_sky_region_gwcs",
-    "test_viewers.py::test_get_viewport_sky_no_wcs",
-    "test_viewers.py::test_get_viewport_pixel_region",
-    "test_viewers.py::test_get_viewport_pixel_region_bad_label",
-    "test_viewers.py::test_get_viewport_pixel_region_no_label",
-     "test_astrowidgets_api.py::TestCenterOffset::test_center_offset_pixel",
-     "test_astrowidgets_api.py::TestCenterOffset::test_center_offset_sky",
-     "test_astrowidgets_api.py::TestCenter::test_center_on_pix",
-     "test_astrowidgets_api.py::TestZoom::test_invalid_zoom_level",
-     "test_astrowidgets_api.py::TestZoom::test_invalid_zoom",
-     "test_astrowidgets_api.py::TestZoom::test_zoom",
-     "test_astrowidgets_api.py::TestCmapStretchCuts::test_colormap_options",
-     "test_astrowidgets_api.py::TestCmapStretchCuts::test_invalid_colormap",
-     "test_astrowidgets_api.py::TestCmapStretchCuts::test_stretch_options",
-     "test_astrowidgets_api.py::TestCmapStretchCuts::test_stretch_astropy",
-     "test_astrowidgets_api.py::TestCmapStretchCuts::test_invalid_stretch",
-     "test_astrowidgets_api.py::TestCmapStretchCuts::test_autocut",
-     "test_astrowidgets_api.py::TestCmapStretchCuts::test_invalid_autocut",
-     "test_astrowidgets_api.py::TestCmapStretchCuts::test_invalid_cuts",
-     "test_astrowidgets_api.py::TestCmapStretchCuts::test_cmap_stretch_cuts",
-     "test_astrowidgets_api.py::TestMarkers::test_invalid_markers",
-     "test_astrowidgets_api.py::TestMarkers::test_mvp_markers",
-     "test_delete_data.py::TestDeleteData::test_reparent_str",
-     "test_delete_data.py::TestDeleteData::test_delete_with_subset_wcs",
-     "test_delete_data.py::TestDeleteWCSLayerWithSubset::test_delete_wcs_layer_with_subset",
-     "test_line_profile_xy.py::TestLineProfileXYPixelLinked::test_plugin",
-     "test_line_profile_xy.py::TestLineProfileXYWCSLinked::test_plugin",
-     "test_linking.py::TestLink_WCS_NoWCS::test_pixel_linking",
-     "test_linking.py::TestLink_WCS_NoWCS::test_wcslink_fallback_pixels",
-     "test_linking.py::TestLink_WCS_FakeWCS::test_pixel_linking",
-     "test_linking.py::TestLink_WCS_FakeWCS::test_badwcs_no_crash",
-     "test_linking.py::TestLink_WCS_WCS::test_pixel_linking",
-     "test_linking.py::TestLink_WCS_WCS::test_wcslink_affine_with_extras",
-     "test_linking.py::TestLink_WCS_WCS::test_wcslink_fullblown",
-     "test_linking.py::TestLink_WCS_WCS::test_invalid_inputs",
-     "test_linking.py::TestLink_WCS_GWCS::test_wcslink_rotated",
-     "test_linking.py::TestLink_GWCS_GWCS::test_pixel_linking",
-     "test_orientation.py::TestDefaultOrientation::test_affine_reset_and_linktype",
-     "test_orientation.py::TestDefaultOrientation::test_astrowidgets_markers_disable_relinking",
-     "test_orientation.py::TestDefaultOrientation::test_markers_plugin_recompute_positions_pixels_to_wcs",
-     "test_orientation.py::TestDefaultOrientation::test_markers_plugin_recompute_positions_wcs_to_pixels",
-     "test_orientation.py::TestNonDefaultOrientation::test_N_up_multi_viewer",
-     "test_orientation.py::TestNonDefaultOrientation::test_custom_orientation",
-     "test_orientation.py::TestDeleteOrientation::test_delete_orientation_multi_viewer",
-     "test_orientation.py::TestDeleteOrientation::test_delete_orientation_with_subset",
-     "test_orientation.py::TestOrientationNoData::test_create_no_data",
-     "test_orientation.py::TestOrientationNoData::test_select_no_data",
-     "test_regions.py::TestLoadRegions::test_regions_invalid",
-     "test_regions.py::TestLoadRegions::test_regions_fully_out_of_bounds",
-     "test_regions.py::TestLoadRegions::test_regions_mask",
-     "test_regions.py::TestLoadRegions::test_regions_pixel",
-     "test_regions.py::TestLoadRegions::test_regions_sky_has_wcs",
-     "test_regions.py::TestLoadRegions::test_regions_annulus_from_load_data",
-     "test_regions.py::TestLoadRegions::test_photutils_pixel",
-     "test_regions.py::TestLoadRegions::test_photutils_sky_has_wcs",
-     "test_regions.py::TestLoadRegionsFromFile::test_ds9_load_all",
-     "test_regions.py::TestLoadRegionsFromFile::test_ds9_load_two_good",
-     "test_regions.py::TestLoadRegionsFromFile::test_ds9_load_one_bad",
-     "test_regions.py::TestGetRegions::test_annulus",
-     "test_simple_aper_phot.py::TestSimpleAperPhot::test_plugin_wcs_dithered",
-     "test_simple_aper_phot.py::TestSimpleAperPhot::test_batch_unpack",
-     "test_simple_aper_phot.py::TestSimpleAperPhot::test_batch_phot",
-     "test_simple_aper_phot.py::TestSimpleAperPhot_NoWCS::test_plugin_no_wcs",
-     "test_simple_aper_phot.py::TestAdvancedAperPhot::test_aperphot",
-     "test_simple_aper_phot.py::TestAdvancedAperPhot::test_sky_background",
-     "test_subset_centroid.py::TestImvizSpatialSubsetCentroidPixelLinked::test_centroiding_pixel",
-     "test_subset_centroid.py::TestImvizSpatialSubsetCentroidWCSLinked::test_centroiding_wcs",
-     "test_tools.py::TestPanZoomTools::test_panzoom_tools",
-     "test_viewer_tools.py::TestContrastBiasTool::test_contrast_bias_mousedrag",
-     "test_viewers.py::TestDeleteData::test_plot_options_after_destroy",
-     "test_viewers.py::TestRegionOverlay::test_add_remove_region_overlay",
-     "test_viewers.py::TestRegionOverlay::test_select_region_overlay",
-     "test_viewers.py::TestRegionOverlay::test_deselect_region_overlay",
-     "test_wcs_utils.py::TestWCSOnly::test_non_wcs_layer_labels"
-]
-
-
-def pytest_collection_modifyitems(config, items):
-    """Mark known-failing cubeviz tests as skipped when running in CI.
-
-    By default this does nothing locally. In CI we detect the standard
-    `CI` environment variable and will skip the curated list of failing tests
-    so that the overall test run can proceed while the failures are being
-    investigated and fixed.
-    """
-
-    # Only skip when running in CI (e.g. GitHub Actions sets CI=true)
-    if os.environ.get("CI", "").lower() not in ("1", "true", "yes"):
-        return
-
-    skip_marker = pytest.mark.skip(reason="Temporarily skipped failing imviz tests in CI")
-    for item in items:
-        # Only consider tests under the cubeviz config directory
-        if "configs/imviz" not in getattr(item, "nodeid", ""):
-            continue
-        for nid_frag in SKIP_TEST_NODEIDS:
-            if nid_frag in item.nodeid:
-                item.add_marker(skip_marker)
-                break
-
-
