+++ jdaviz/configs/rampviz/conftest.py
@@ -47,5 +47,3 @@ def pytest_collection_modifyitems(config, items):
             if nid_frag in item.nodeid:
                 item.add_marker(skip_marker)
                 break
-
-
